app/requirements.txt
flask==2.3.3
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
redis==5.0.1

